Use TAP automake driver when possible.
With automake 1.12, it's possible to use a custom test driver, and so we
can use its TAP driver instead. When the sass-spec framework is set up
to output TAP, then the automake driver will understand all the subtests
and be able to count them individually.

Note, automake 1.11 does not support custom test drivers, so I added a
check in configure to see if the automake-produced Makefile.in includes
the TEST_LOG_DRIVER variable. This seems to be the only way to test for
the custom log driver ability. Once the automake requirements are bumped
to 1.12, this test should be removed (especially if some newer automake
stops defining this variable).