Add automake-1.12 tests on Travis.
This requires installing a PPA to get it, but it's worth it to be able
to use the TAP driver with automake. Hopefully, Travis will upgrade to
the next LTS and get a newer automake within a year, anyway.

Also, I disabled testing of automake-1.12+clang, since that's basically
the same as automake-1.11+clang.