expose list-delim option for simplifying huge css changes
After a while when projects grow larger changes to the CSS pose
significant changes in the Git history. I.e., as hundreds of CSS
elements are combined into single elements, changing, adding, or
removing just a single item will modify the entire CSS line.

As an example, imagine the following example:

  a, b, c, d {
    block: none;
  }

Now we define a new CSS rule, "e", and we end up with:

  a, b, c, d, e {
    block: none;
  }

This is all fine in a small project, but we have lots of cases where the
identifier list ("a, b, c, d, e") extends hundreds and hundreds of
elements, turning this into huge Git commit diffs.

This proposal turns the above into the following:

  a,
  b,
  c,
  d {
    block: none;
  }

And when adding "e", it simply changes to:

  a,
  b,
  c,
  d,
  e {
    block: none;
  }

As you can imagine this will greatly simplify the Git commit changes and
therefore also the commit size and especially (!) merge conflicts.
E.g., when merging two branches, both of which added/modified an element
*somewhere* in the identifier list, things get ugly fast.

Now I hear you say "don't commit generated .css files", but we live in a
world where people want to be able to develop on projects without having
to compile the sass/scss into css manually. This can only be done if the
intermediate versions (rather than the release versions) of the CSS code
are available without being required to install node.js / etc.