Replace the GitHub with a managed heroku proxy
We sometimes hit the GitHub API rate limit. To get around this we
have a hosted reverse proxy that has a baked in auth token. We
occasionally have issue with the proxy going down, and can only wait
for it to be bumped.

This PR uses a proxy deployed on heroku with a very restrictive auth
token. This should put an end to the proxy going down, and make it
possible for multiple people to manage the server if needed.