Avoid modifying Expression in ast_node_to_sass_value
Previously, the argument was defined as `const Expression_Ptr`.

The `const` there was useless, as it didn't mean the constness of Expression but merely that `val` pointer could not be reassigned within the function (due to how pointer typedefs and consts interact).

Assuming the intended argument type was `const Expression_Ptr *`, I've changed it to `Expression_Ptr_Const`.

This required making Color conversion functions `const` and always copying the result.
This likely doesn't matter much, especially since color arithmetic is deprecated.

The `disp` property is no longer clear when copying, instead it is cleared when a setter is called.