travis: Update to xenial, disable LTO
For clang, travis sets `gold` as the default linker but doesn't have `LLVMgold.so` installed.
For gcc, there are warnings from `ranlib`, although the build succeeds.

    ranlib: base64vlq.o: plugin needed to handle lto object

This commit disables LTO on Travis, and also fixes up makefiles.

Also updates Travis distro from trusty (14.04) to xenial (16.04), which
comes with gcc5 so we don't have to install it via apt.