Simplify string concatenation
Simply do `a + b + c`.

In C++ 11, if at least one of the operands is an rvalue reference, it'll
be basically as good as `a += b; b += c;`. Besides, all of these strings
are short.

Ideally we'd use string views and something like `absl::StrCat`
throughout, but we don't have anything like that here.