Implement `capture_stderr` C-API option
As an embeddable library we should try to play nice with
our parents resources. It is therefore desirable to have
a way to capture messages on stderr. This makes it easier
and safer to use libsass in eg. threaded environments.

Core errors (mainly memory allocation problems) and
dev/debug data is still printed to stderr directly.