Remove version.h from source directory
The version.h should be generated from version.h.in. If the build
directory is out of the source directory, the version.h is generated in
build directory and the original version.h in source directory is not
overwritten. However, the latter is used during the build, resulting in
a missing runtime version:

$ sassc --version
sassc: [NA]
libsass: [NA]
sass2scss: 1.1.1
sass: 3.5

Remove version.h from source directory to ensure that the version.h in
build directory is used during the build.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>